+++ MonsterChase/Player.h
@@ -9,6 +9,6 @@ public:
 	~Player();
 
 	void MovePlayer(char move, Player* player);
-	void MovePlayerMain(char movement, Player* player, const int playX, const int playY);
+	void MovePlayerMain(char movement, Player* player, const float playX, const float playY);
 };
